Summary
-------

Apply a variant of Kahan summation that keeps a running total for the cumulative sum and another for the cumulative fractional errors.  Start the accumulation with 1.0 to establish the invariant that ``|csum| > |x|``.  This makes sure that no information is thrown-away in the computation of the adjustment to *frac*.

The improved summation routine makes hypot() and dist() fully commutative in most cases.

Measurements
------------

$ # baseline
$ py test_hypot_commutativity.py
Counter({1: 999905, 2: 95})
$ py test_hypot_commutativity.py
Counter({1: 999916, 2: 84})
$ py test_hypot_commutativity.py
Counter({1: 999876, 2: 124})
$ py test_hypot_commutativity.py
Counter({1: 999885, 2: 115})

$ # patched
$ py test_hypot_commutativity.py
Counter({1: 1000000})
$ py test_hypot_commutativity.py
Counter({1: 1000000})
$ py test_hypot_commutativity.py
Counter({1: 1000000})
$ py test_hypot_commutativity.py
Counter({1: 1000000})


Test script
----------

<pre>
# test_hypot_commutativity.py
from math import hypot
from random import expovariate
from itertools import permutations, starmap
from collections import Counter

trials = 1_000_000
n = 5

lambdas = [1.0E-3 ** exp for exp in range(n)]
print(Counter(len(set(starmap(hypot, permutations(map(expovariate, lambdas))))) for i in range(trials)))
</pre>

Timings
-------

The patch also gives a negligible performance boost:

<pre>
$ # baseline
$ py -m timeit -r9 -s 'from math import hypot' 'hypot(15.0, 14.0, 13.0, 12.0, 11.0, 10.0, 9.0, 8.0, 7.0, 6.0, 5.0, 4.0, 3.0, 2.0, 1.0)'
2000000 loops, best of 9: 181 nsec per loop
$ py -m timeit -r9 -s 'from math import hypot' 'hypot(10.0, 9.0, 8.0, 7.0, 6.0, 5.0, 4.0, 3.0, 2.0, 1.0)'
2000000 loops, best of 9: 136 nsec per loop
$ py -m timeit -r9 -s 'from math import hypot' 'hypot(5.0, 4.0, 3.0, 2.0, 1.0)'
5000000 loops, best of 9: 94 nsec per loop

$ # patched
$ py -m timeit -r9 -s 'from math import hypot' 'hypot(15.0, 14.0, 13.0, 12.0, 11.0, 10.0, 9.0, 8.0, 7.0, 6.0, 5.0, 4.0, 3.0, 2.0, 1.0)'
2000000 loops, best of 9: 179 nsec per loop
$ py -m timeit -r9 -s 'from math import hypot' 'hypot(10.0, 9.0, 8.0, 7.0, 6.0, 5.0, 4.0, 3.0, 2.0, 1.0)'
2000000 loops, best of 9: 134 nsec per loop
$ py -m timeit -r9 -s 'from math import hypot' 'hypot(5.0, 4.0, 3.0, 2.0, 1.0)'
5000000 loops, best of 9: 92.5 nsec per loop
</pre>